Рефакторинг редьюсера
